AI-628: Consolidate generated content admin listings
