+++ app/controllers/api/admin/goods_nomenclature_self_texts_controller.rb
@@ -1,79 +1,20 @@
 module Api
   module Admin
     class GoodsNomenclatureSelfTextsController < AdminController
-      def index
-        render json: serialized_collection
-      end
+      include GeneratedContentListing
 
       private
 
-      def serialized_collection
-        GoodsNomenclatures::GoodsNomenclatureSelfTextSerializer.new(
-          paginated_dataset.all,
-          is_collection: true,
-          meta: pagination_meta,
-        ).serializable_hash
+      def model_class
+        GoodsNomenclatureSelfText
       end
 
-      def pagination_meta
-        {
-          pagination: {
-            page: current_page,
-            per_page:,
-            total_count: paginated_dataset.pagination_record_count,
-          },
-        }
+      def serializer_class
+        GoodsNomenclatures::GoodsNomenclatureSelfTextSerializer
       end
 
-      def paginated_dataset
-        @paginated_dataset ||= filtered_dataset.paginate(current_page, per_page)
-      end
-
-      def filtered_dataset
-        dataset = GoodsNomenclatureSelfText
-          .admin_listing(include_expired: expired_listing?)
-          .search(params[:q])
-          .for_nomenclature_type(params[:type])
-          .for_status(params[:status])
-          .for_score_category(params[:score_category])
-
-        dataset = dataset.exclude(Sequel[:goods_nomenclature_self_texts][:approved] => true) if normal_review_listing?
-
-        apply_sorting(dataset)
-      end
-
-      def normal_review_listing?
-        params[:status].blank? && !search_query?
-      end
-
-      def expired_listing?
-        params[:status] == 'expired'
-      end
-
-      def search_query?
-        q = params[:q].to_s.strip
-
-        q.match?(/\A\d{2,10}\z/) || q.length >= 2
-      end
-
-      def apply_sorting(dataset)
-        col = sort_column
-        dir = sort_direction == 'desc' ? :desc : :asc
-
-        dataset.order(Sequel.public_send(dir, col, nulls: :last))
-      end
-
-      def sort_column
-        allowed = {
-          'score' => Sequel.lit('score'),
-          'goods_nomenclature_item_id' => Sequel[:goods_nomenclature_self_texts][:goods_nomenclature_item_id],
-        }
-
-        allowed[params[:sort]] || Sequel.lit('score')
-      end
-
-      def sort_direction
-        %w[asc desc].include?(params[:direction]) ? params[:direction] : 'asc'
+      def listing_table
+        Sequel[:goods_nomenclature_self_texts]
       end
     end
   end
